backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
python-multipart==0.0.20
PyPDF2==3.0.1
python-docx==1.1.2
openpyxl==3.1.5
